+++ lib/lerk/user_accounts.rb
@@ -29,12 +29,20 @@ module Lerk
       @bot.command(
         :link,
         description: 'Tell the bot about your other (eg Steam) accounts',
-        usage: '!link (--steam) (steam_id)',
-        min_args: 2,
+        usage: '!link [--steam steam_id]',
+        min_args: 0,
         max_args: 2,
-      ) do  |event, type, arg|
+      ) do  |event, *args|
         return if Util.ignored?(event)
-        command_link(event, type, arg)
+
+        if args.length == 0
+          command_link_connections(event)
+        elsif args.length == 2
+          type, arg = args
+          command_link_manual(event, type, arg)
+        else
+          return 'Invalid usage. Check `!help link`'
+        end
       end
 
       @bot.command(
@@ -49,8 +57,8 @@ module Lerk
       end
     end
 
-    def self.command_link(event, type, arg)
-      @logger.command(event, 'link', { type: type, arg: arg })
+    def self.command_link_manual(event, type, arg)
+      @logger.command(event, 'link', { target: 'manual', type: type, arg: arg })
 
       account_type = account_type(type)
       unless account_type
@@ -66,10 +74,24 @@ module Lerk
           return "Unknown Steam ID: #{ arg }"
         end
 
-        user.update(steam_account_id: account_id)
+        user.update(steam_account_id: account_id, steam_account_verified: false)
         @cmd_link_counter.increment({ status: :success }, event: event)
         return "Linked Steam account #{ account_id } to Discord user #{ user.discord_id }."
       end
+    end
+
+    def self.command_link_connections(event)
+      @logger.command(event, 'link', { target: 'connections' })
+
+      user = Util.discord_user_from_database(event)
+      url = user.oauth_auth_url
+      dm_msg = "Please click on the following link to link your Discord accounts:\n#{ url }"
+      event.user.pm(dm_msg)
+
+      return "#{ event.user.mention }: You have been sent a direct message with a link. Open it to allow me reading your Discord-connected services."
+
+    rescue Discordrb::Errors::NoPermission => e
+      return "I was unable to send you a direct message, please adjust your settings to allow receiving DMs:\nRight click the server icon -> Privacy Settings -> Allow direct messages from server members"
     end
 
     def self.command_unlink(event, type)
@@ -84,7 +106,7 @@ module Lerk
       user = Util.discord_user_from_database(event)
       if account_type == 'steam'
         @cmd_unlink_counter.increment({ status: :success }, event: event)
-        user.update(steam_account_id: nil)
+        user.update(steam_account_id: nil, steam_account_verified: false)
         return "Unlinked Steam account from Discord user #{ user.discord_id }."
       end
     end